Change DOS_Drive::FileOpen functions to take uint8_t for flags
This sets the flags variable that was changed to uint8_t in a previous commit.
This function gets called from DOS_OpenFile which passes it a uint8_t already.

+++ include/dos_system.h
@@ -345,7 +345,7 @@ public:
 	DOS_Drive();
 	virtual ~DOS_Drive() = default;
 
-	virtual bool FileOpen(DOS_File** file, char* name, uint32_t flags) = 0;
+	virtual bool FileOpen(DOS_File** file, char* name, uint8_t flags) = 0;
 	virtual bool FileCreate(DOS_File** file, char* name,
 	                        FatAttributeFlags attributes) = 0;
 	virtual bool FileUnlink(char* _name)=0;
