+++ src/dos/drive_iso.cpp
@@ -208,7 +208,7 @@ void isoDrive::Activate(void) {
 	UpdateMscdex(driveLetter, fileName, subUnit);
 }
 
-bool isoDrive::FileOpen(DOS_File **file, char *name, uint32_t flags) {
+bool isoDrive::FileOpen(DOS_File **file, char *name, uint8_t flags) {
 	if ((flags & 0x0f) == OPEN_WRITE) {
 		DOS_SetError(DOSERR_ACCESS_DENIED);
 		return false;
